Fix the bug Wolvzrain encountered
Idk about the things that lead up to this state, but if `data.ascStatToggles` is `undefined`, the use of the `||` on the new line 217 means we try to reference something on `undefined` leading to a crash. There are other ways to fix this issue, like a second null check or using a negated, DeMorgan's-d check (`!(player.ascStatToggles[4] !== undefined && data.ascStatToggles[4] !== undefined)`) to avoid trying to reference a property of `undefined`. I took the just initialize `data.ascStatToggles` approach, but I have no context. I think you might also want to set `data.ascStatToggles[4]` in the if branch, too?