transition to new save exporter/history
![](https://i.imgur.com/jir6A3B.png)

TODO list:
- [ ] Move from localStorage to IndexedDB.
- [ ] Move export/import buttons to save history (depending on feedback)
- [ ] Increase number of saves (depending on feedback) - right now it's set as 10.
- [ ] Options for auto saving, rather than manually clicking the button (user defined interval).

As I'm busy with finals, I'm not sure when I'll get around to finishing this PR up. So, I'll let it sit here and see if anyone wants to check some features off.

(I tried keeping the style similar to the history tab, so thanks Kewne for the html/css I stole from you)